Modified http channels module Added documentation Fixed some coding issues I found in the original ex_ari code Added variables parameter support to the originate method Added originate_with_id method Added variables parameter support to the create method Added hangup map to the hangup method. Allows specifying a reason code and a reason.
Changed several methods to use a map as opposed to always specify fields.
This allows more flexibility and will work better with the way Asterisk defines the ARI interface.
Also if Asterisk adds new fields for the methods, we do not have to modify ex_ari for every field.
To help developers understand what to put in the map, the additional documentation
Specifies the parameters supported, whether any are required, range, and some defaults.

Added support for rtp_stats method.
Added support for external_media method.